squash: Address review comments.
