+++ react/features/video-quality/middleware.js
@@ -2,19 +2,12 @@
 
 import { CONFERENCE_JOINED } from '../base/conference';
 import { SET_CONFIG } from '../base/config';
-import { getParticipantCount } from '../base/participants';
-import { MiddlewareRegistry, StateListenerRegistry } from '../base/redux';
-import { shouldDisplayTileView } from '../video-layout';
+import { MiddlewareRegistry } from '../base/redux';
 
-import { setPreferredVideoQuality, setMaxReceiverVideoQuality } from './actions';
-import { VIDEO_QUALITY_LEVELS } from './constants';
-import { getReceiverVideoQualityLevel } from './functions';
+import { setPreferredVideoQuality } from './actions';
 import logger from './logger';
-import { getMinHeightForQualityLvlMap } from './selector';
 
 import './subscriber';
-
-declare var APP: Object;
 
 /**
  * Implements the middleware of the feature video-quality.
@@ -52,110 +45,3 @@ MiddlewareRegistry.register(({ dispatch, getState }) => next => action => {
 
     return result;
 });
-
-/**
- * Implements a state listener in order to calculate max receiver video quality.
- */
-StateListenerRegistry.register(
-    /* selector */ state => {
-        const { reducedUI } = state['features/base/responsive-ui'];
-        const _shouldDisplayTileView = shouldDisplayTileView(state);
-        const thumbnailSize = state['features/filmstrip']?.tileViewDimensions?.thumbnailSize;
-        const participantCount = getParticipantCount(state);
-
-        return {
-            displayTileView: _shouldDisplayTileView,
-            participantCount,
-            reducedUI,
-            thumbnailHeight: thumbnailSize?.height
-        };
-    },
-    /* listener */ ({ displayTileView, participantCount, reducedUI, thumbnailHeight }, { dispatch, getState }) => {
-        const state = getState();
-        const { maxReceiverVideoQuality } = state['features/video-quality'];
-        const { maxFullResolutionParticipants = 2 } = state['features/base/config'];
-
-        let newMaxRecvVideoQuality = VIDEO_QUALITY_LEVELS.ULTRA;
-
-        if (reducedUI) {
-            newMaxRecvVideoQuality = VIDEO_QUALITY_LEVELS.LOW;
-        } else if (displayTileView && !Number.isNaN(thumbnailHeight)) {
-            newMaxRecvVideoQuality = getReceiverVideoQualityLevel(thumbnailHeight, getMinHeightForQualityLvlMap(state));
-
-            // Override HD level calculated for the thumbnail height when # of participants threshold is exceeded
-            if (maxReceiverVideoQuality !== newMaxRecvVideoQuality && maxFullResolutionParticipants !== -1) {
-                const override
-                    = participantCount > maxFullResolutionParticipants
-                        && newMaxRecvVideoQuality > VIDEO_QUALITY_LEVELS.STANDARD;
-
-                logger.info(`Video quality level for thumbnail height: ${thumbnailHeight}, `
-                    + `is: ${newMaxRecvVideoQuality}, `
-                    + `override: ${String(override)}, `
-                    + `max full res N: ${maxFullResolutionParticipants}`);
-
-                if (override) {
-                    newMaxRecvVideoQuality = VIDEO_QUALITY_LEVELS.STANDARD;
-                }
-            }
-        }
-
-        if (maxReceiverVideoQuality !== newMaxRecvVideoQuality) {
-            dispatch(setMaxReceiverVideoQuality(newMaxRecvVideoQuality));
-        }
-    }, {
-        deepEquals: true
-    });
-
-/**
- * Helper function for updating the preferred sender video constraint, based
- * on the user preference.
- *
- * @param {JitsiConference} conference - The JitsiConference instance for the
- * current call.
- * @param {number} preferred - The user preferred max frame height.
- * @returns {void}
- */
-function _setSenderVideoConstraint(conference, preferred) {
-    if (conference) {
-        conference.setSenderVideoConstraint(preferred)
-            .catch(err => {
-                logger.error(`Changing sender resolution to ${preferred} failed - ${err} `);
-            });
-    }
-}
-
-/**
- * Registers a change handler for state['features/base/conference'] to update
- * the preferred video quality levels based on user preferred and internal
- * settings.
- */
-StateListenerRegistry.register(
-    /* selector */ state => {
-        const { conference } = state['features/base/conference'];
-        const {
-            maxReceiverVideoQuality,
-            preferredVideoQuality
-        } = state['features/video-quality'];
-
-        return {
-            conference,
-            maxReceiverVideoQuality,
-            preferredVideoQuality
-        };
-    },
-    /* listener */ (currentState, store, previousState = {}) => {
-        const {
-            conference,
-            preferredVideoQuality
-        } = currentState;
-        const changedConference = conference !== previousState.conference;
-        const changedPreferredVideoQuality = preferredVideoQuality !== previousState.preferredVideoQuality;
-
-        if (changedConference || changedPreferredVideoQuality) {
-            _setSenderVideoConstraint(conference, preferredVideoQuality);
-        }
-
-        if (typeof APP !== 'undefined' && changedPreferredVideoQuality) {
-            APP.API.notifyVideoQualityChanged(preferredVideoQuality);
-        }
-    });
